separated birthdaybar elements into individual files
