+++ birthdaytracker/lib/widgets/birthday_bar_picture.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class BdarBarPicture extends StatelessWidget {
+  BdarBarPicture({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return Container(
+      height: 80,
+      width: 80,
+      child: Center(
+          child: Container(
+        height: 70,
+        width: 70,
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(100), color: Colors.white),
+      )),
+    );
+  }
+}
